feat(mxl-player-components): add initialization options to customize playlist
The following new values can be set during initialization: You can set the mutability of the playlist, provide optional DiscovererInfo, set the default repeat mode, display the index of each file in the playlist, and mark a specific entry as playing. This last option can be used to create a copy of the playlist during playback.